(parser) Use valid size for iteration variables
This commit does not change anything for the moment, because of Orcc issue
explained in orcc/orcc#86. The IR produced reused the first i variable
declared as int(size=32).

Once the mentionned issue will be fixed and merged in the master,
3 different variables will be declared for the 3 foreach blocks. This
will cause crash (segfault).

The bug described here can be reproduced by renaming i variables into
something unused in the procedure.